add composition to ConstructorPage
