+++ Project/Fixtures/fixturePages.ts
@@ -9,7 +9,9 @@ import { SearchResultsPage } from '../Pages/SearchResultsPage';
 import { ComparisonPage } from '../Pages/ComparisonPage';
 import { HeaderComponent } from '../Components/headerComponent';
 import { ProfilePage } from '../Pages/ProfilePage';
-import { ConstructorPage } from '../Pages/ConstructorPage';
+import { ConstructorPage, FAQComponent, ConsultationModal } from '../Pages/ConstructorPage';
+import { ComputerConfiguratorPage } from '../Pages/ComputerConfiguratorPage';
+
 
 type Pages = {
     catalogPage: CatalogPage;
@@ -22,6 +24,9 @@ type Pages = {
     comparisonPage: ComparisonPage;
     profilePage: ProfilePage;
     constructorPage: ConstructorPage;
+    faqComponent: FAQComponent;
+    consultationModal: ConsultationModal;
+    computerConfiguratorPage: ComputerConfiguratorPage;
 }
 
 type Components = {
@@ -73,5 +78,17 @@ export const test = base.extend<Pages & Components> ({
     constructorPage: ({page}, use) => {
         const constructorPage = new ConstructorPage(page);
         use(constructorPage);
+    },
+    faqComponent: ({page}, use) => {
+        const faqComponent = new FAQComponent(page);
+        use(faqComponent);
+    },
+    consultationModal: ({page}, use) => {
+        const consultationModal = new ConsultationModal(page);
+        use(consultationModal);
+    },
+    computerConfiguratorPage: ({page}, use) => {
+        const computerConfiguratorPage = new ComputerConfiguratorPage(page);
+        use(computerConfiguratorPage);
     }
 })
