+++ tests/constructor/buildPC.spec.ts
@@ -1,18 +1,19 @@
 import { test } from "../../Project/Fixtures/fixturePages";
 import {
   callReason,
-  headerTextContent,
-  componentsSections,
   constructorButton,
   buildPCButton,
-  buyButton,
   getConsultationButton,
-  assembledPCs
 } from "../../Project/Pages/ConstructorPage";
+import {
+  headerTextContent,
+  componentsSections,
+  buyButton,
+  assembledPCs,
+} from "../../Project/Pages/ComputerConfiguratorPage";
 import { headerText } from "../../Project/Pages/FilteringResultsPage";
 
 test.describe("Build PC page", () => {
-
   const modalFormData = ["Test", "0000000000", "10000", callReason.newPC];
   const sectionTitles = [
     "Послуги",
@@ -39,6 +40,7 @@ test.describe("Build PC page", () => {
   test("Build PC page: Get consultation", async ({
     homePage,
     constructorPage,
+    consultationModal,
   }) => {
     await homePage.navigateToBaseURL();
     await homePage.confirmCityModal();
@@ -47,7 +49,7 @@ test.describe("Build PC page", () => {
     await constructorPage.clickButton(buildPCButton);
     await constructorPage.checkHeader(headerTextContent.configurePC);
     await constructorPage.clickButton(getConsultationButton);
-    await constructorPage.getConsultation(modalFormData);
+    await consultationModal.getConsultation(modalFormData);
   });
 
   test("Build PC page: assembled PCs", async ({
@@ -55,13 +57,14 @@ test.describe("Build PC page", () => {
     constructorPage,
     filteringResultsPage,
     catalogPage,
+    computerConfiguratorPage,
   }) => {
     await homePage.navigateToBaseURL();
     await homePage.confirmCityModal();
     await homePage.clickButton(constructorButton);
 
     await constructorPage.clickButton(buildPCButton);
-    await constructorPage.clickButton(assembledPCs);
+    await computerConfiguratorPage.clickButton(assembledPCs);
     await filteringResultsPage.checkHeader(headerText.PCs);
     await catalogPage.checkURL("https://telemart.ua/ua/pc/");
   });
@@ -69,6 +72,7 @@ test.describe("Build PC page", () => {
   test("Build PC page: Configure PC", async ({
     homePage,
     constructorPage,
+    computerConfiguratorPage,
     checkoutPage,
   }) => {
     await homePage.navigateToBaseURL();
@@ -76,14 +80,14 @@ test.describe("Build PC page", () => {
     await homePage.clickButton(constructorButton);
 
     await constructorPage.clickButton(buildPCButton);
-    await constructorPage.checkNumberOfSections(4);
-    await constructorPage.checkNumberOfItems(35);
-    await constructorPage.checkSectionTitles(sectionTitles);
+    await computerConfiguratorPage.checkNumberOfSections(4);
+    await computerConfiguratorPage.checkNumberOfItems(35);
+    await computerConfiguratorPage.checkSectionTitles(sectionTitles);
 
-    await constructorPage.addComponents(componentsTitles);
-    await constructorPage.checkConstructionCompleted();
-    await constructorPage.clickButton(buyButton);
-    await constructorPage.submitBuyModal();
+    await computerConfiguratorPage.addComponents(componentsTitles);
+    await computerConfiguratorPage.checkConstructionCompleted();
+    await computerConfiguratorPage.clickButton(buyButton);
+    await computerConfiguratorPage.submitBuyModal();
 
     await checkoutPage.checkURL("https://telemart.ua/ua/order/");
     await checkoutPage.fillSurname("Druzhchenko");
